+++ src/app/app.component.ts
@@ -1,7 +1,11 @@
 import { Router, RouterModule } from "@angular/router";
 import { AfterViewInit, Component, ElementRef,  ViewChild } from '@angular/core';
+import { Store } from "@ngrx/store";
 
-import {AppSettingsService} from "./core/services/app-settings.service";
+import { AppSettingsService } from "./core/services/app-settings.service";
+
+import * as ProductsActions from './core/@ngrx/products';
+import * as RouterActions from './core/@ngrx/router';
 
 import { HeaderComponent } from "./core/components/header/header.component";
 
@@ -16,7 +20,8 @@ import { HeaderComponent } from "./core/components/header/header.component";
 export class AppComponent implements AfterViewInit {
   @ViewChild('appTitle') appTitle!: ElementRef<HTMLHeadingElement>;
 
-  constructor(private router: Router, private appSettingsService: AppSettingsService) {
+  constructor(private router: Router, private appSettingsService: AppSettingsService, private store: Store) {
+    this.store.dispatch(ProductsActions.getProducts());
     this.appSettingsService.initSettings();
     const replacer = (key: string, value: any): string =>
       typeof value === 'function' ? value.name : value;
@@ -28,6 +33,6 @@ export class AppComponent implements AfterViewInit {
   }
 
   onGoHome() {
-    this.router.navigate(['/']);
+    this.store.dispatch(RouterActions.goHome());
   }
 }