+++ src/app/products/components/product-component/product.component.ts
@@ -1,4 +1,4 @@
-import {Component, EventEmitter, Input, Output} from '@angular/core';
+import { Component, EventEmitter, Input, OnChanges, Output } from '@angular/core';
 
 import { IProductModel } from "../../models/product.model";
 
@@ -7,11 +7,19 @@ import { IProductModel } from "../../models/product.model";
   templateUrl: './product.component.html',
   styleUrls: ['./product.component.scss']
 })
-export class ProductComponent {
+export class ProductComponent implements OnChanges {
   @Input() product!: IProductModel;
-  @Output() addProductToCart = new EventEmitter<IProductModel>()
+  @Input() cartIds!: ReadonlyArray<number>;
+
+  @Output() addProductToCart = new EventEmitter<IProductModel>();
+
+  inCart = false;
+
+  ngOnChanges() {
+    this.inCart = this.cartIds?.includes(this.product.id);
+  }
 
   onAddToCart() {
-    this.addProductToCart.emit(this.product)
+    this.addProductToCart.emit(this.product);
   }
 }